rename: 이벤트와 호스트 대시보드 각 기능에 맞게 파일 분리
