+++ src/features/dashboard/hook/useEventHook.ts
@@ -1,22 +1,7 @@
-import { useMutation, useQuery } from '@tanstack/react-query';
+import { useMutation } from '@tanstack/react-query';
 import { useParams } from 'react-router-dom';
-import { getHostDashboard, updateEventInfo } from '../api/event';
 import { UpdateEventRequest } from '../model/event';
-import { HostDashboardResponse } from '../../../entities/host/model/hostDashboard';
-
-export const useGetEventHook = () => {
-  const { id } = useParams();
-
-  const eventId = Number(id);
-
-  const { data: eventInfo } = useQuery<HostDashboardResponse>({
-    queryKey: ['eventInfo', eventId],
-    queryFn: () => getHostDashboard(eventId),
-  });
-
-  return { eventInfo };
-};
-
+import { updateEventInfo } from '../api/event';
 export const useUpdateEventHook = () => {
   const { id } = useParams();
   const eventId = Number(id);
